@@ -14,42 +14,41 @@
     "ci": "npm run lint && npm run test:once"
   },
   "dependencies": {
-    "@angular/animations": "8.2.1",
-    "@angular/common": "8.2.1",
-    "@angular/compiler": "8.2.1",
-    "@angular/core": "8.2.1",
-    "@angular/forms": "8.2.1",
-    "@angular/platform-browser": "8.2.1",
-    "@angular/platform-browser-dynamic": "8.2.1",
-    "@angular/router": "8.2.1",
+    "@angular/animations": "11.2.7",
+    "@angular/common": "11.2.7",
+    "@angular/compiler": "11.2.7",
+    "@angular/core": "11.2.7",
+    "@angular/forms": "11.2.7",
+    "@angular/platform-browser": "11.2.7",
+    "@angular/platform-browser-dynamic": "11.2.7",
+    "@angular/router": "11.2.7",
     "core-js": "2.6.3",
-    "rxjs": "6.5.2",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.0",
-    "rxjs-compat": "6.5.2"
+    "rxjs": "6.6.6",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.802.1",
-    "@angular/cli": "8.2.1",
-    "@angular/compiler-cli": "8.2.1",
-    "@angular/language-service": "8.2.1",
-    "@types/chrome": "0.0.77",
-    "@types/jasmine": "3.3.7",
-    "@types/jasminewd2": "2.0.6",
-    "@types/node": "10.12.18",
+    "@angular-devkit/build-angular": "~0.1102.6",
+    "@angular/cli": "11.2.6",
+    "@angular/compiler-cli": "11.2.7",
+    "@angular/language-service": "11.2.7",
+    "@types/chrome": "0.0.133",
+    "@types/jasmine": "3.6.9",
+    "@types/jasminewd2": "2.0.8",
+    "@types/node": "14.14.37",
     "@types/gulp": "4.0.5",
-    "codelyzer": "^5.0.1",
-    "jasmine-core": "3.3.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "3.1.4",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "2.0.4",
-    "karma-jasmine": "2.0.1",
-    "karma-jasmine-html-reporter": "1.4.0",
-    "protractor": "5.4.2",
-    "ts-node": "8.0.1",
-    "tslint": "5.12.1",
-    "typescript": "3.5.3",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "9.1.1",
+    "tslint": "6.1.3",
+    "typescript": "4.1.5",
     "gulp": "4.0.0",
     "gulp-replace": "1.0.0",
     "gulp-zip": "4.2.0",
