@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { AppComponent } from './app.component';
 import { Router } from '@angular/router';
@@ -7,7 +7,7 @@ describe('AppComponent', () => {
   let component: AppComponent;
   let fixture: ComponentFixture<AppComponent>;
   let mockWindow;
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     mockWindow = {
       $registerUiWindow: jasmine.createSpy('$registerUiWindow').and.stub()
     };
