@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { Router } from '@angular/router';
 import { IoProxy } from 'app/common/ioproxy';
 import { InstallComponent } from './install.component';
@@ -6,7 +6,7 @@ import { InstallComponent } from './install.component';
 describe('InstallComponent', () => {
   let component: InstallComponent;
   let fixture: ComponentFixture<InstallComponent>;
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     const mockRouter = {},
       mockIoProxy = {};
     TestBed.configureTestingModule({
